fix: make a separate file for figure blot
